+++ BetterBulldozer/ChangeLog.md
@@ -2,3 +2,4 @@
 * Added Dependency for Unified Icon Library
 * Fixed Raycasting Lanes
 * Added Raycasting Icons
+* Added utilities networks to raycast marker networks.